# pkg/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Atelino 后端 API 文档",
        "title": "Atelino API",
        "contact": {},
        "version": "1.0.8"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/hitokoto": {
            "get": {
                "description": "从数据库中随机获取一条一言记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "一言"
                ],
                "summary": "随机获取一言",
                "operationId": "getHitokotoRandom",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/atelino_internal_dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/atelino_internal_dto.HitokotoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "没有找到对应的一言",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一条新的一言记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "一言"
                ],
                "summary": "添加一言",
                "operationId": "createHitokoto",
                "parameters": [
                    {
                        "description": "一言内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.CreateHitokotoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/atelino_internal_dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/atelino_internal_dto.HitokotoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "409": {
                        "description": "该一言已存在",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/hitokoto/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从数据库中查询所有的一言记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "一言"
                ],
                "summary": "获取一言列表",
                "operationId": "getHitokotoList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/atelino_internal_dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/atelino_internal_dto.HitokotoResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/hitokoto/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "传入一言的 ID，从数据库中查询对应的记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "一言"
                ],
                "summary": "获取一言",
                "operationId": "getHitokotoByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "一言 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/atelino_internal_dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/atelino_internal_dto.HitokotoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "404": {
                        "description": "没有找到对应的一言",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "传入一言的 ID，从数据库中删除对应的记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "一言"
                ],
                "summary": "删除一言",
                "operationId": "deleteHitokotoByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "一言 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "404": {
                        "description": "没有找到对应的一言",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    },
                    "500": {
                        "description": "数据库错误",
                        "schema": {
                            "$ref": "#/definitions/atelino_internal_dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "atelino_internal_dto.CreateHitokotoRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "atelino_internal_dto.HitokotoResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "atelino_internal_dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}